requirements.txt
streamlit>=1.22.0
pdf2image>=1.16.3
pydantic>=2.0.0
pandas>=1.5.0
Pillow>=9.0.0
requests>=2.28.0
numpy>=1.22.0
python-dotenv>=1.0.0
mistralai==0.0.13
